readdir/readdirplus now synthesize . and .. in listings as appropriate
- touched a bit of extra code because i decided to make the DirentCollector
  return the estimated entry count as a signed int, instead of the
  implicitly unsigned size_t
- fuse_readdir and fuse_readdirplus became more complicated, in exactly
  the same way. so i turned them into lightweight wrappers around
  fuse_readdir_common

+++ src/generic/void_inflight_action.h
@@ -35,11 +35,11 @@ public:
     explicit DirentCollector(const DirentCollectorSpec &spec)
         : plus_mode_(spec.plus_mode), max_entries_(spec.max_entries) {}
 
-    [[nodiscard]] size_t estimate_remaining_entries() const {
+    [[nodiscard]] int estimate_remaining_entries() const {
       if (count_ >= max_entries_) {
         return 0;
       }
-      return max_entries_ - count_;
+      return static_cast<int>(max_entries_ - count_);
     }
 
     [[nodiscard]] DirentAddResult try_add(std::string_view,
